Fixes #3 and #102 issues in prost-dev repo.
This is a small bug fix on top of commit d40a9e4.  That commit (d40a9e4)
tried to fix excessive Prost! memory consumption by not storing 
(i.e. erasing) the genomic locations of reads that exceed 
`max_locations_allowed`.  However, that commit introduced a bug.

Basically, the bug is as follows: We have to erase the genomic locations
in designation_step_one.  However, if a sequence with too many gen_locs
has a designation of DESIGNATION_ONE_OR_TWO, then it would be parsed
in designation_step_two.  However, for designation_step_two to work,
all gen_locs for all sequences need to be kept, so that 
designation_step_two can figure out who is a ‘1’ and who is a ‘2’.  
Of course though, we aren’t storing the gen_locs for all sequences 
anymore (i.e. for sets with len(gen_locs) > max_locations_allowed).

Solution is to simply re-designate such sequences with a new
designation (DESIGNATION_MAX_ALLOWED_SEQUENCES_EXCEEDED) and call it 
a day.

Now such sequences are reported as ‘MLAE’ (maximum locations allowed
exceeded).

Tested, it’s good.