Fixed reverse annotation bug.
Basically, it was incorrectly not including a reverse annotation if
there was a forward annotation in another species.

This behavior may have not been consistent oddly.  But definitely
the logic that I fixed today was incorrect, and running Prost a
second time seemed to fix the original bug report by Thomas.