Manuscript resources + Small bug fixed: corrected two iso_* columns.
Added files to manuscript_resources/ that were used for the Prost runs
in the manuscript.

The iso_snp_seed and iso_snp_central_offset were always set to FALSE.
This is corrected now.

Additionally, added empty __slots__ to GenLocBin and SeedBin to ensure
a bug like this doesn’t sneak in again.