Minor definition change of is_supplementary_edited / iso_snp_supp.
This came up while fixing up the other iso_* columns.  Realized
there were cases that we really should not be calling
"is_supplementary_edited".  For example, 16=4X is most likely to be
untemplated addition of 4 NTs (as opposed to NT17 in the supplementary
region being edited and then the untemplated addition of 3 NTs).
Updated code to reflect this.